# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Radhiyan Pet and Care",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/datastaticlocation": {
            "delete": {
                "tags": [
                    "Data Static"
                ],
                "summary": "Delete Data Static Location",
                "description": "Delete data static here , need id data static",
                "operationId": "datastaticlocation",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {},
                            "examples": {
                                "delete id 1": {
                                    "summary": "Delete data static",
                                    "value": {
                                        "id": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Delete data static Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Delete data static Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/datastatic": {
            "get": {
                "tags": [
                    "Data Static"
                ],
                "summary": "Get Data Static",
                "description": "Get Data Static",
                "operationId": "datastatic",
                "parameters": [
                    {
                        "name": "body",
                        "in": "path",
                        "required": true,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "orderby": {
                                            "type": "text",
                                            "example": "asc"
                                        },
                                        "column": {
                                            "type": "text",
                                            "example": "value"
                                        },
                                        "keyword": {
                                            "type": "text",
                                            "example": "Messenger"
                                        },
                                        "page": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "total_per_page": {
                                            "type": "number",
                                            "example": "14"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Get Data Static Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Get Data Static Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Insert Data Static",
                "description": "Insert Data Static",
                "operationId": "Insert Data Static",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "keyword",
                                    "name"
                                ],
                                "properties": {
                                    "keyword": {
                                        "type": "text"
                                    },
                                    "name": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {},
                            "examples": {
                                "Insert Data Static": {
                                    "summary": "Insert Data static",
                                    "value": {
                                        "keyword": "Pemakaian",
                                        "name": "Sharing Account"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}